+++ test/support/test_rpc_module.ex
@@ -15,6 +15,14 @@ defmodule Ethers.TestRPCModule do
     {:ok, opts[:tx_hash] || "tx_hash"}
   end
 
+  def eth_send_raw_transaction(params, opts) do
+    if pid = opts[:send_params_to_pid] do
+      send(pid, params)
+    end
+
+    {:ok, opts[:tx_hash] || "tx_hash"}
+  end
+
   def eth_call(params, block, opts) do
     if pid = opts[:send_back_to_pid] do
       send(pid, :eth_call)