+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Support [EIP-1191](https://eips.ethereum.org/EIPS/eip-1191): Add chain id to mixed-case checksum address encoding
 - Add [EIP-4844](https://eips.ethereum.org/EIPS/eip-4844) transaction support
 - Add [EIP-2930](https://eips.ethereum.org/EIPS/eip-2930) transaction support
+- Add `Ethers.blob_base_fee/1` to calculate blob fee per gas
+- Make default gas and fee margin configurable
+
+### Bug Fixes
+
+- Fix encoding of transactions without any input data
+- Encode `access_list` and `blob_versioned_hashed` correctly in RPC transaction requests
 
 ## v0.6.1 (2025-01-02)